requirements.txt
flask_sqlalchemy
flask_security_too
flask_mail
flask_dotenv
flask_migrate